Fix String

+++ codebook/String/Manacher.cpp
@@ -2,7 +2,7 @@
 vector <int> manacher(string tmp) {
   string s = "&";
   for (char c : tmp) s.pb(c), s.pb('%');
-  int l = 0, r = 0, n = s.size();
+  int l = 0, r = 0, n = (int)s.size();
   vector <int> Z(n);
   for (int i = 0; i < n; ++i) {
     Z[i] = r > i ? min(Z[2 * l - i], r - i) : 1;
